refactor: move modules to modules folder
