+++ src/modules/auth/dto/register-response.dto.ts
