+++ src/modules/student/student.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { StudentService } from './student.service';
 import { StudentController } from './student.controller';
 import { JwtModule } from '@nestjs/jwt';
-import { JwtStrategy } from 'src/auth/strategy';
+import { JwtStrategy } from 'src/modules/auth/strategy';
 
 @Module({
   imports: [JwtModule.register({})],
