+++ src/modules/teacher/dto/index.ts
@@ -1,4 +1,4 @@
-export * from '../../common/dto/answer.dto';
+export * from '../../../common/dto/answer.dto';
 export * from './question.dto';
 export * from './quiz.dto';
 export * from './question-response.dto';
